refactor(db): extract message reader
